Moves services to their own directory
