+++ src/services/contract-service.js
@@ -1,10 +1,10 @@
-import ClaimHolderRegisteredContract from "./../contracts/build/contracts/ClaimHolderRegistered.json"
-import ClaimHolderPresignedContract from "./../contracts/build/contracts/ClaimHolderPresigned.json"
-import ListingsRegistryContract from "./../contracts/build/contracts/ListingsRegistry.json"
-import ListingContract from "./../contracts/build/contracts/Listing.json"
-import PurchaseContract from "./../contracts/build/contracts/Purchase.json"
-import UserRegistryContract from "./../contracts/build/contracts/UserRegistry.json"
-import OriginIdentityContract from "./../contracts/build/contracts/OriginIdentity.json"
+import ClaimHolderRegisteredContract from "./../../contracts/build/contracts/ClaimHolderRegistered.json"
+import ClaimHolderPresignedContract from "./../../contracts/build/contracts/ClaimHolderPresigned.json"
+import ListingsRegistryContract from "./../../contracts/build/contracts/ListingsRegistry.json"
+import ListingContract from "./../../contracts/build/contracts/Listing.json"
+import PurchaseContract from "./../../contracts/build/contracts/Purchase.json"
+import UserRegistryContract from "./../../contracts/build/contracts/UserRegistry.json"
+import OriginIdentityContract from "./../../contracts/build/contracts/OriginIdentity.json"
 import bs58 from "bs58"
 import Web3 from "web3"
